an idea
Creation of a warehouse class so all the coordinated to be intialized by that and it will make the accessing of this easier as well as their initialization by the user... it also works as a link between the model and the gui. There rest of the changes are expirementatiosns with the cost estimation strategy. If this will work then we are step a closer fot the a* to be implemented.